mudanças

+++ server/auth/controller/TechnicianController/TechOTP.js
@@ -102,12 +102,13 @@ exports.confirmOTP = async (req, res) => {
         };
 
         const validOTP = await bcrypt.compare(otp, OTPHash);
-
-        await Technician.updateOne({ cod_technician: Id }, { verified: true });
-
-        await EmailOTP.deleteOne({ userId: Id });
-
-        return res.status(200).json({ validOTP });
+        if (validOTP) {
+            await Technician.updateOne({ _id: Id }, { verified: true });
+            await EmailOTP.deleteOne({ userId: Id });
+            return res.status(200).json({ msg: 'OTP Confirmado com sucesso!' });
+        } else {
+            return res.status(422).json({ msg: 'OTP Inválido!' });
+        };
     } catch (error) {
         console.error(error)
         res.status(500).json({ msg: error });
